@@ -1,11 +1,10 @@
-// swift-tools-version: 5.10
+// swift-tools-version: 6.0
 
 import PackageDescription
 
 let settings: [SwiftSetting] = [
     .unsafeFlags(["-warnings-as-errors"]),
-    .enableExperimentalFeature("ExistentialAny"),
-    .enableExperimentalFeature("StrictConcurrency=targeted")
+    .enableExperimentalFeature("ExistentialAny")
 ]
 
 let package = Package(
@@ -58,5 +57,6 @@ let package = Package(
                            .product(name: "MEGADomainMock", package: "MEGADomain")
                           ],
             swiftSettings: settings)
-    ]
+    ],
+    swiftLanguageModes: [.v6]
 )
