@@ -1,7 +1,7 @@
 import MEGAAnalyticsDomain
 import MEGAAnalyticsiOS
 
-final class ViewIdProviderAdapter: ViewIdProvider {
+final class ViewIdProviderAdapter: ViewIdProvider, Sendable {
     private let viewIdUseCase: any ViewIDUseCaseProtocol
     
     init(viewIdUseCase: some ViewIDUseCaseProtocol) {
