@@ -92,7 +92,8 @@ private func makeNullViewModel() -> VideoListViewModel {
         fileSearchUseCase: Preview_FilesSearchUseCase(),
         thumbnailLoader: Preview_ThumbnailLoader(),
         sensitiveNodeUseCase: Preview_SensitiveNodeUseCase(),
-        nodeUseCase: Preview_NodeUseCase()
+        nodeUseCase: Preview_NodeUseCase(),
+        featureFlagProvider: Preview_FeatureFlagProvider(isFeatureFlagEnabled: false)
     )
 }
 
